Refactor sendReminder function and add new testindex.js file
